Changed 'pages' to 'app'
