+++ src/app/api/feedback/route.js
@@ -1,0 +1,18 @@
+import { submitFeedback } from "../../services/feedback";
+
+export async function POST(req) {
+    const { email, feedback } = await req.json();
+    const result = await submitFeedback(email, feedback);
+
+    if (!result) {
+        return Response.json({
+            status: 500,
+            message: "Internal server error",
+        });
+    } else {
+        return Response.json({
+            status: 200,
+            message: "OK",
+        });
+    }
+}
