+++ src/app/api/search/route.js
@@ -1,0 +1,18 @@
+import { getSearchResult } from "../../services/search";
+
+export async function POST(req) {
+    const { keyword } = await req.json();
+    const result = await getSearchResult(keyword);
+    if (result == null) {
+        return Response.json({
+            status: 404,
+            message: "Not found",
+        });
+    } else {
+        return Response.json({
+            status: 200,
+            message: "OK",
+            data: result,
+        });
+    }
+}
