+++ src/app/layout.js
@@ -1,0 +1,17 @@
+import { Inter } from "next/font/google";
+import "./globals.css";
+
+const inter = Inter({ subsets: ["latin"] });
+
+export const metadata = {
+    title: "Thuật ngữ lịch sử",
+    description: "Tra cứu thuật ngữ lịch sử",
+};
+
+export default function RootLayout({ children }) {
+    return (
+        <html lang="en">
+            <body className={inter.className}>{children}</body>
+        </html>
+    );
+}
