+++ src/app/page.js
@@ -1,0 +1,8 @@
+"use client";
+
+import { redirect } from "next/navigation";
+
+export default function Home() {
+    redirect("/home");
+}
+
