+++ src/app/services/filter.js
@@ -1,0 +1,13 @@
+import { removeTone } from "../utils/utils";
+import { sql } from "@vercel/postgres";
+
+export async function getFilterResult(query) {
+    console.log(query);
+    query = removeTone(query).replaceAll(/[\s-]/g, "");
+    const result = await sql`
+        SELECT * FROM Keywords 
+        WHERE keywordCode 
+        LIKE ${"%" + query + "%"}
+        LIMIT 6`;
+    return result.rows;
+}
