+++ src/app/utils/utils.js
@@ -1,0 +1,8 @@
+export function removeTone(str) {
+    if (typeof str !== "string") return "";
+    return str
+        .toLowerCase()
+        .replace("đ", "d")
+        .normalize("NFD")
+        .replace(/[\u0300-\u036f]/g, "");
+}
