--- src/pages/index.js
@@ -1,9 +1,0 @@
-import Body from "./components";
-
-export default function Home() {
-    return (
-        <div className="wrap bg-[#FED38F]">
-            <Body />
-        </div>
-    );
-}
